+++ packages/core/di/index.ts
@@ -7,4 +7,4 @@ export * from './self';
 export * from './skip-self';
 export * from './optional';
 export * from './host';
-// export * from './inject';
+export * from './inject';