+++ packages/core/di/metadata.ts
@@ -7,3 +7,5 @@ export const metadataOfHost = 'nvMetadataOfHost';
 export const metadataOfSelf = 'nvMetadataOfSelf';
 
 export const metadataOfSkipSelf = 'nvMetadataOfSkipSelf';
+
+export const metadataOfInject = 'nvMetadataOfInject';